refactored gen back to toplevel
